Moved the code for hiding the desktop icons to the shell class and also simplified it.
