+++ src/Kirei.Infrastructure/DesktopAPI/Desktop.cs
@@ -1,31 +1,21 @@
 ﻿using Kirei.Application;
-using Kirei.Infrastructure.Native;
-using Kirei.Infrastructure.Native.Enums;
-
-using System;
 
 namespace Kirei.Infrastructure.DesktopAPI
 {
     public class Desktop :
         IDesktop
     {
-        private const string DefView_WINDOW_NAME = "SHELLDLL_DefView";
-        private const string WorkerW_CLASS_NAME = "WorkerW";
-
-        private readonly IntPtr _command = new IntPtr(0x7402);
-
         private readonly Shell _shell;
         private readonly TaskBar _taskBar;
-        private readonly IntPtr _shellDefViewHandle;
 
         private bool windowsMinimized = false;
         private bool taskBarHidden = false;
+        private bool desktopIconsHidden = false;
 
         public Desktop()
         {
             _shell = new Shell();
             _taskBar = new TaskBar();
-            _shellDefViewHandle = FetchShellDefViewHandle();
         }
 
         public void ToggleTaskBar()
@@ -40,11 +30,12 @@ namespace Kirei.Infrastructure.DesktopAPI
 
         public void ToggleIcons()
         {
-            User32.SendMessage(
-                _shellDefViewHandle,
-                WindowsMessages.WM_COMMAND,
-                _command,
-                IntPtr.Zero);
+            if (desktopIconsHidden)
+                _shell.ShowDesktopIcons();
+            else
+                _shell.HideDesktopIcons();
+
+            desktopIconsHidden = !taskBarHidden;
         }
 
         public void ToggleWindows()
@@ -55,31 +46,6 @@ namespace Kirei.Infrastructure.DesktopAPI
                 _shell.MinimizeWindows();
 
             windowsMinimized = !windowsMinimized;
-        }
-
-        private IntPtr FetchShellDefViewHandle()
-        {
-            var shellWindowHandle = User32.GetShellWindow();
-            IntPtr handle = default;
-
-            User32.EnumWindows(delegate (IntPtr hWnd, IntPtr lParam)
-            {
-                if (!User32.GetClassName(hWnd).Equals(WorkerW_CLASS_NAME, StringComparison.OrdinalIgnoreCase))
-                    return true;
-
-                handle = User32.FindWindowEx(
-                    hWnd,
-                    IntPtr.Zero,
-                    DefView_WINDOW_NAME,
-                    IntPtr.Zero);
-
-                if (handle == IntPtr.Zero)
-                    return true;
-
-                return false;
-            }, IntPtr.Zero);
-
-            return handle;
         }
     }
 }
